requirements.txt
# Compatibility with python 3.9 version
pymongo==3.6
dnspython==2.4.2
python-dotenv-1.0.0
icecream